Datepicker update
Email for reference:

Hey Mitch,
Small update with the flatpickr config, can you please change it to:

$('.datepicker').flatpickr({format: "Y-m-d",altFormat: "d M
Y",altInput: true});})

There was an issue where PHP wasn't receiving the date in the correct
format, I think possibly from mobile when the input is type=date, as it
reverts to the local format, but PHP was expecting it in "d M Y" format.

So I've remove the type=date from the input, and just added a class
'datepicker' to the input. And also using the format "Y-m-d" which
seems to get around the errors I was encountering.